requirements.txt
# Core LLM & Orchestration
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
langchain-chroma==0.1.4
openai==1.40.0

# Vector Store
chromadb==0.5.5

# Database
sqlalchemy==2.0.32
aiosqlite==0.20.0

# API
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0

# Document Loading
pypdf==4.3.1
unstructured==0.15.0
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
rich==13.7.1
tiktoken==0.7.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0

# Notebooks
jupyter==1.0.0
ipykernel==6.29.5
